requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
typing-extensions==4.12.2

